# openapi.yaml
openapi: 3.0.3
info:
  title: EduFlow API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/workspaces/onboarding-state:
    get:
      summary: Get the current workspace onboarding state
      operationId: getWorkspaceOnboardingState
      description: Returns routing state derived from existing workspace, subscription-access, and payment-request records.
      security: [{ betterAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
      responses:
        '200':
          description: Current onboarding state
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { $ref: '#/components/schemas/WorkspaceOnboardingState' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/WorkspaceMembershipRequired' }
  /api/v1/workspaces/onboard:
    post:
      summary: Create the workspace for an approved account purchase
      operationId: createWorkspaceOnboard
      description: Consumes the authenticated user's approved unused subscription purchase atomically, then creates the active workspace, Owner membership, default settings, and attached subscription. A user who already belongs to a workspace cannot create another one.
      security: [{ betterAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkspaceOnboard' }
      responses:
        '201':
          description: Active workspace created from the approved purchase
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { $ref: '#/components/schemas/OnboardedWorkspace' } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409':
          {
            description: Workspace already exists,
            slug already exists,
            or creation is not unlocked by an approved purchase,
          }
  /api/v1/workspaces/dashboard-summary:
    get:
      summary: Get the current workspace dashboard summary
      operationId: getWorkspaceDashboardSummary
      description: Returns the authenticated member's own workspace identity, access state, current plan dates, effective feature entitlements, membership count, latest subscription-payment status, and tenant-scoped operational aggregates. Financial values are exact decimal strings. The workspace is resolved only from the verified X-Workspace-Id membership context.
      security: [{ betterAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - name: month
          in: query
          required: false
          description: First day of the month to aggregate, in YYYY-MM-01 format. Defaults to the current Bangladesh month.
          schema: { type: string, format: date, pattern: "^[1-9]\\d{3}-\\d{2}-01$" }
        - name: batchId
          in: query
          required: false
          description: Restricts scheduled-batch, attendance, fee-collection, and outstanding-fee aggregates to one batch in the current workspace. Salary and expense totals remain workspace-wide because those records are not batch-owned.
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Current workspace dashboard information
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { $ref: '#/components/schemas/WorkspaceDashboardSummary' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/WorkspaceMembershipRequired' }
        '404': { description: Workspace not found }
...
  /api/v1/workspaces/{id}:
    get:
      summary: Get a Platform Owner workspace detail
      operationId: getPlatformWorkspaceDetail
      security: [{ betterAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/WorkspacePathId' }
      responses:
        '200':
          {
            description: Workspace detail,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        required: [data],
                        properties: { data: { $ref: '#/components/schemas/WorkspaceDetail' } },
                      },
                  },
              },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
        '404': { description: Workspace not found }
  /api/v1/workspaces:
    get:
      summary: List workspaces for the Platform Owner
      operationId: listPlatformWorkspaces
      security:
        - betterAuth: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
        - { name: search, in: query, schema: { type: string, minLength: 1, maxLength: 150 } }
        - {
            name: workspaceStatus,
            in: query,
            schema:
              {
                type: string,
                enum: [pending, active, locked, suspended, scheduled_deletion, deleted],
              },
          }
        - {
            name: subscriptionStatus,
            in: query,
            schema:
              {
                type: string,
                enum: [trial, pending, active, renewal_due, expired, cancelled, suspended],
              },
          }
        - {
            name: accessStatus,
            in: query,
            schema:
              {
                type: string,
                enum:
                  [
                    verification_pending,
                    payment_pending,
                    trial,
                    active,
                    renewal_due,
                    subscription_expired,
                    locked,
                    scheduled_for_deletion,
                    suspended,
                  ],
              },
          }
        - name: lifecycleQueue
          in: query
          description: Restrict the list to locked and scheduled-for-deletion workspaces for the Platform Owner deletion queue.
          schema: { type: boolean }
      responses:
        '200':
          description: Paginated workspace listing with raw workspace, subscription, and effective access summaries
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/WorkspaceSummary' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
...
components:
  securitySchemes:
    betterAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
...
    WorkspacePathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    WorkspaceId:
      name: X-Workspace-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    WorkspaceOnboardingState:
      type: object
      required: [step, workspaceStatus, paymentPending, access]
      properties:
        step:
          { type: string, enum: [workspace_created, subscription_required, payment_pending, ready] }
        workspaceStatus:
          { type: string, enum: [pending, active, locked, suspended, scheduled_deletion, deleted] }
        paymentPending: { type: boolean }
        access:
          type: object
          required: [allowed, status, reason]
          properties:
            allowed: { type: boolean }
            status:
              {
                type: string,
                enum:
                  [
                    verification_pending,
                    payment_pending,
                    trial,
                    active,
                    renewal_due,
                    subscription_expired,
                    locked,
                    scheduled_for_deletion,
                    suspended,
                  ],
              }
            reason: { type: string }
    CreateWorkspaceOnboard:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 150 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 100 }
        phone: { type: string, pattern: "^(?:\\+8801\\d{9}|01\\d{9})$" }
        email: { type: string, format: email, maxLength: 255 }
        address: { type: string, minLength: 1, maxLength: 2000 }
    OnboardedWorkspace:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        address: { type: string, nullable: true }
        status: { type: string, enum: [pending] }
        createdAt: { type: string, format: date-time }
...
    WorkspaceDetail:
      type: object
      required:
        [
          workspace,
          subscription,
          access,
          subscriptionHistory,
          recentPayments,
          activeEntitlementOverrides,
        ]
      properties:
        workspace:
          type: object
          required: [id, name, slug, status, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            name: { type: string, nullable: true }
            slug: { type: string, nullable: true }
            phone: { type: string, nullable: true }
            email: { type: string, format: email, nullable: true }
            address: { type: string, nullable: true }
            status:
              {
                type: string,
                enum: [pending, active, locked, suspended, scheduled_deletion, deleted],
              }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            activatedAt: { type: string, format: date-time, nullable: true }
            lockedAt: { type: string, format: date-time, nullable: true }
            scheduledDeleteAt: { type: string, format: date-time, nullable: true }
        subscription: { nullable: true, type: object }
        access:
          {
            type: object,
            required: [allowed, status, reason],
            properties:
              { allowed: { type: boolean }, status: { type: string }, reason: { type: string } },
          }
        subscriptionHistory: { type: array, items: { type: object } }
        recentPayments: { type: array, items: { $ref: '#/components/schemas/PaymentRequest' } }
        activeEntitlementOverrides: { type: array, items: { type: object } }
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
    WorkspaceSummary:
      type: object
      required: [id, name, slug, workspaceStatus, createdAt, subscription, access]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        slug: { type: string, nullable: true }
        workspaceStatus:
          { type: string, enum: [pending, active, locked, suspended, scheduled_deletion, deleted] }
        createdAt: { type: string, format: date-time }
        lockedAt: { type: string, format: date-time, nullable: true }
        scheduledDeleteAt: { type: string, format: date-time, nullable: true }
        subscription:
          nullable: true
          type: object
          properties:
            id: { type: string, format: uuid }
            status:
              {
                type: string,
                enum: [trial, pending, active, renewal_due, expired, cancelled, suspended],
              }
            startsAt: { type: string, format: date-time, nullable: true }
            expiresAt: { type: string, format: date-time, nullable: true }
            trialEndsAt: { type: string, format: date-time, nullable: true }
            plan:
              {
                type: object,
                nullable: true,
                properties:
                  {
                    id: { type: string, format: uuid },
                    name: { type: string },
                    slug: { type: string },
                  },
              }
        access:
          type: object
          required: [allowed, status, reason]
          properties:
            allowed: { type: boolean }
            status: { type: string }
            reason: { type: string }
    WorkspaceDashboardSummary:
      type: object
      required:
        [workspace, access, subscription, memberCount, entitlements, latestPayment, operational]
      properties:
        workspace:
          type: object
          required: [id, name, status]
          properties:
            id: { type: string, format: uuid }
            name: { type: string, nullable: true }
            status:
              {
                type: string,
                enum: [pending, active, locked, suspended, scheduled_deletion, deleted],
              }
        access:
          type: object
          required: [allowed, status, reason]
          properties:
            allowed: { type: boolean }
            status:
              {
                type: string,
                enum:
                  [
                    verification_pending,
                    payment_pending,
                    trial,
                    active,
                    renewal_due,
                    subscription_expired,
                    locked,
                    scheduled_for_deletion,
                    suspended,
                  ],
              }
            reason: { type: string }
        subscription:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/SubscriptionDashboardStatus' }]
        memberCount: { type: integer, minimum: 0 }
        entitlements:
          type: array
          items:
            type: object
            required: [key, enabled, limit]
            properties:
              key: { type: string }
              enabled: { type: boolean }
              limit:
                { type: string, nullable: true, description: Integer limit serialized as a string. }
        latestPayment:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/PaymentDashboardStatus' }]
        operational:
          type: object
          required: [activeStudents, activeBatches, activeTeachers, today, monthlyFinance]
          properties:
            activeStudents: { type: integer, minimum: 0 }
            activeBatches: { type: integer, minimum: 0 }
            activeTeachers: { type: integer, minimum: 0 }
            today:
              type: object
              required: [date, scheduledBatchCount, attendance]
              properties:
                date: { type: string, format: date, description: Current date in Asia/Dhaka. }
                scheduledBatchCount: { type: integer, minimum: 0 }
                attendance:
                  type: object
                  required:
                    [
                      sessionCount,
                      finalizedSessionCount,
                      draftSessionCount,
                      presentCount,
                      absentCount,
                    ]
                  properties:
                    sessionCount: { type: integer, minimum: 0 }
                    finalizedSessionCount: { type: integer, minimum: 0 }
                    draftSessionCount: { type: integer, minimum: 0 }
                    presentCount: { type: integer, minimum: 0 }
                    absentCount: { type: integer, minimum: 0 }
            monthlyFinance:
              type: object
              required: [month, collectedFees, outstandingFees, paidSalaries, expenses, netCashFlow]
              properties:
                month: { type: string, format: date }
                collectedFees: { type: string, pattern: '^\\d+\\.\\d{2}$' }
                outstandingFees: { type: string, pattern: '^\\d+\\.\\d{2}$' }
                paidSalaries: { type: string, pattern: '^\\d+\\.\\d{2}$' }
                expenses: { type: string, pattern: '^\\d+\\.\\d{2}$' }
                netCashFlow: { type: string, pattern: '^-?\\d+\\.\\d{2}$' }
    SubscriptionDashboardStatus:
      type: object
      required: [status, planName, startsAt, expiresAt, trialEndsAt, renewalDueAt]
      properties:
        status:
          {
            type: string,
            enum: [trial, pending, active, renewal_due, expired, cancelled, suspended],
          }
        planName: { type: string, nullable: true }
        startsAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        trialEndsAt: { type: string, format: date-time, nullable: true }
        renewalDueAt: { type: string, format: date-time, nullable: true }
    PaymentDashboardStatus:
      type: object
      required: [status, createdAt, reviewedAt]
      properties:
        status: { type: string, enum: [pending, approved, rejected] }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time, nullable: true }
...
    PaymentRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        planId: { type: string, format: uuid, nullable: true }
        amountMinor: { type: string, pattern: '^\\d+$' }
        paymentMethod: { type: string, enum: [cash, bkash, nagad, rocket, other] }
        senderNumber: { type: string }
        transactionId: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        reviewedAt: { type: string, format: date-time, nullable: true }
        reviewedByUserId: { type: string, nullable: true }
        rejectionReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
  responses:
...
    ValidationError:
      description: Invalid request
    Unauthenticated:
      description: Authentication required
    WorkspaceMembershipRequired:
      description: Workspace membership required
    PlatformOwnerRequired:
      description: Platform Owner authorization required
...